Use double for auction prices and bids
Replaced float-based monetary values with double across the auctions flow to improve numeric precision and consistency. This updates entities, DTOs, service and controller method signatures, currency exchange rate handling (including Optional type), and seeded article price literals in the initializer.

+++ src/main/java/es/deusto/sd/auctions/entity/Article.java
@@ -79,11 +79,11 @@ public class Article {
 		this.title = title;
 	}
 
-	public float getInitialPrice() {
+	public double getInitialPrice() {
 		return initialPrice;
 	}
 
-	public void setInitialPrice(float initialPrice) {
+	public void setInitialPrice(double initialPrice) {
 		this.initialPrice = initialPrice;
 	}
